fix: Move postcode helper to common
