+++ cfl_common/common/models.py
@@ -1,14 +1,14 @@
-import pgeocode
 import re
 from datetime import timedelta
 from uuid import uuid4
 
+import pgeocode
 from django.contrib.auth.models import User
 from django.db import models
 from django.utils import timezone
 from django_countries.fields import CountryField
 
-from portal.helpers.organisation import sanitise_uk_postcode
+from .helpers.organisation import sanitise_uk_postcode
 
 
 class UserProfile(models.Model):
